Implement basic Plex playback info
Shows media that is currently played via the configured Plex server just
like the Kodi implementation did before. Unfortunately Plex has no way
to determine if current playback is going to the living room stereo or
somewhere entirely different.
That would need to be determined via other means. (For a start, maybe
check if the playing plex client is considered to be on the local
network by Plex. That would already solve the 'listening to music at
work' case.)